Committing progress so far

+++ crates/chia-consensus/src/consensus_constants.rs
@@ -17,7 +17,7 @@ pub struct ConsensusConstants {
     pub max_sub_slot_blocks: u32,
 
     /// The number of signage points per sub-slot (including the 0th sp at the sub-slot start).
-    pub num_sps_sub_slot: u32,
+    pub num_sps_sub_slot: u8,
 
     /// The sub_slot_iters for the first epoch.
     pub sub_slot_iters_starting: u64,
